All puts and print removed from game logic and test files deleted
