--- main-test.rb
@@ -1,95 +1,0 @@
-require_relative 'game-logic-test'
-
-puts "\n" + '*' * 50
-puts "\n          ---- Tic Tac TRONIX 2000 ----"
-puts "\n        An strategy game for two players."
-puts "\n             Creative Commons 2020"
-puts "\n" + '*' * 50
-puts "\n Tic Tac TRONIX 2000 is based on the classic Tic
- Tac Toe game."
-
-puts "\n INSTRUCTIONS: to play select \"Y\", you'll be
- asked to introduce the name of both Players. Once
- set up, players will take turns to leave their
- marks (X's or O's) on the board. The player who
- first has 3 marks in a row wins the game."
-
-puts "\n                   __RULES__
-\n1. The game is played on a grid that's 3x3 squares.
-2. Player One is X's and Player Two is 0's.
-3. X always starts the game.
-4. Players take turns putting their marks in unselected
-   squares.
-5. The first player to get 3 marks in a row (up,
-   down, across, or diagonally) is the winner.
-6. When all 9 squares are full, the game is over.
-   If no player has 3 marks in a row, the game ends
-   in a tie."
-
-prompt = '> '
-
-print "\nDo you wanna play? Y/n: "
-answer = gets.chomp
-
-while !/^y(es){0,1}$/i.match(answer) || !/^no{0,1}$/i.match(answer)
-  if /^y(es){0,1}$/i.match(answer)
-    prompt = '> '
-    puts "\nPlayer 1 choose your name: "
-    print prompt
-    player1 = Player.new(gets.chomp.upcase)
-
-    puts "\nNow, Player 2 choose your name: "
-    print prompt
-    player2 = Player.new(gets.chomp.upcase)
-
-    game = Game.new(player1, player2, answer)
-
-    puts "\nWelcome players!"
-    puts "\n#{player1.name} you'll play as the '#{player1.mark}' and
-    #{player2.name} you'll play as the '#{player2.mark}'."
-
-    puts "\nThis is the board\n"
-
-    board = Board.new
-
-    def print_board(game_board)
-      game_board.each_with_index do |x, y|
-        separator = [1, 2, 4, 5, 7, 8]
-        print "  #{game_board[y]} "
-        print '|' if separator.include?(y + 1)
-        print "\n" if ((y + 1).to_i % 3).zero?
-      end
-    end
-
-    print_board(board.availible_spaces)
-    
-    until game.result
-      puts "\n#{game.active_player.name} pick a number to put your game piece"
-      pick = gets.chomp.to_i
-      board.space_selection(pick, game.active_player.mark, game.active_player.choices)
-      print_board(board.availible_spaces)
-      if board.win_validation(game.active_player.choices)
-        puts "Congratulations #{game.active_player.name}, you're a winner"
-        break
-      elsif board.draw_validation
-        puts "It's a draw"
-        break
-      end
-      game.switch_players
-    end
-
-    break
-
-  elsif /^no{0,1}$/i.match(answer)
-
-    puts 'Goodbye...'
-    break
-
-  else
-
-    puts "\nDidn't get you. You said you wanna play: Yes or No?"
-    print prompt
-    answer = gets.chomp
-
-  end
-end
